+++ utils/azone2/s3d.h
@@ -1,7 +1,7 @@
 #ifndef __EQCLIENT_S3D_H_
 #define __EQCLIENT_S3D_H_
 #include <stdio.h>
-#include "types.h"
+#include "../../common/types.h"
 
 typedef struct struct_header {
   uint32 offset;
@@ -31,7 +31,7 @@ typedef struct struct_fn_entry {
 
 typedef struct s3d_object {
   FILE *fp;
-  long count;
+  int32 count;
   char **filenames;
   uint32 *files;
 } s3d_object;
@@ -40,7 +40,7 @@ typedef struct s3d_object {
 extern "C" {
 #endif
 
-int S3D_Init(s3d_object *obj, FILE *fp);
+int32 S3D_Init(s3d_object *obj, FILE *fp);
 size_t S3D_GetFile(s3d_object *obj, char *filename, uchar **out);
 
 #ifdef __cplusplus
